src/tests/requirements.txt
starlette
uvicorn
kubernetes
google-auth
google-auth-oauthlib
google-cloud-firestore 
firebase-admin 
httpx
google-auth
